Actor for performing rollups for provided metrics against KairosDB.

+++ app/com/arpnetworking/kairos/client/models/Sampling.java
@@ -30,7 +30,7 @@ public final class Sampling {
         return _unit;
     }
 
-    public int getValue() {
+    public long getValue() {
         return _value;
     }
